Tech | MW 0.17.27 (#2097)

+++ data/src/main/java/com/anytypeio/anytype/data/auth/repo/block/BlockRemote.kt
@@ -156,8 +156,8 @@ interface BlockRemote {
         sorts: List<DVSort>,
         filters: List<DVFilter>,
         keys: List<String>,
-        offset: Int,
-        limit: Int,
+        offset: Long,
+        limit: Long,
         beforeId: Id?,
         afterId: Id?,
     ): SearchResult
