+++ domain/src/main/java/com/anytypeio/anytype/domain/block/repo/BlockRepository.kt
@@ -203,8 +203,8 @@ interface BlockRepository {
         sorts: List<DVSort>,
         filters: List<DVFilter>,
         keys: List<String>,
-        offset: Int,
-        limit: Int,
+        offset: Long,
+        limit: Long,
         beforeId: Id?,
         afterId: Id?,
     ): SearchResult
