+++ domain/src/main/java/com/anytypeio/anytype/domain/search/SearchObjectsWithSubscription.kt
@@ -30,15 +30,15 @@ class SearchObjectsWithSubscription(
         val filters: List<DVFilter> = emptyList(),
         val fulltext: String = EMPTY_TEXT,
         val keys: List<String>,
-        val offset: Int = INIT_OFFSET,
-        val limit: Int = LIMIT,
+        val offset: Long = INIT_OFFSET,
+        val limit: Long = LIMIT,
         val beforeId: Id?,
         val afterId: Id?,
     )
 
     companion object {
         const val EMPTY_TEXT = ""
-        const val LIMIT = 1000
-        const val INIT_OFFSET = 0
+        const val LIMIT = 1000L
+        const val INIT_OFFSET = 0L
     }
 }
