+++ middleware/src/main/java/com/anytypeio/anytype/middleware/block/BlockMiddleware.kt
@@ -371,8 +371,8 @@ class BlockMiddleware(
         sorts: List<DVSort>,
         filters: List<DVFilter>,
         keys: List<String>,
-        offset: Int,
-        limit: Int,
+        offset: Long,
+        limit: Long,
         beforeId: Id?,
         afterId: Id?
     ): SearchResult = middleware.searchObjectsWithSubscription(
