+++ middleware/src/main/java/com/anytypeio/anytype/middleware/interactor/Middleware.kt
@@ -1224,8 +1224,8 @@ class Middleware(
         sorts: List<DVSort>,
         filters: List<DVFilter>,
         keys: List<String>,
-        offset: Int,
-        limit: Int,
+        offset: Long,
+        limit: Long,
         beforeId: Id?,
         afterId: Id?,
     ): SearchResult {
